Move bootstraps from lrsystem to algorithms
